+++ 06_puppy/mohankrishna/map_store.go
@@ -2,36 +2,6 @@ package main
 
 import "fmt"
 
-// Error codes
-const (
-	ErrInvalidInput int = iota
-	ErrDuplicate
-	ErrNotFound
-)
-
-type Puppy struct {
-	ID     uint32
-	Breed  string
-	Colour string
-	Value  int
-}
-
-type Storer interface {
-	CreatePuppy(puppy *Puppy) error
-	ReadPuppy(ID uint32) (*Puppy, error)
-	UpdatePuppy(ID uint32, puppy *Puppy) error
-	DeletePuppy(ID uint32) (bool, error)
-}
-
-type Error struct {
-	Code    int
-	Message string
-}
-
-func (e *Error) Error() string {
-	return e.Message
-}
-
 type MapStore struct {
 	m map[uint32]*Puppy
 }
